Recomendacion completa

+++ backEnd/src/main/java/dev/EduPrep/eduprep/entities/StudentCourseGrade.java
@@ -46,7 +46,7 @@ public class StudentCourseGrade {
         this.course = course;
     }
 
-    public float getGrade() {
+    public double getGrade() {
         return grade;
     }
